New AG-feature: the ability to manually specify dependencies between attributes. Example:
----------------------------------------------
DATA Root | Root  a : Aap

DATA Aap | Een

SEM Aap [ | | p : Int q : Int ]
  | Een  lhs.p = 1
         lhs.q = 2

SEM Root [ | | z : Int ]
  | Root
      lhs.z = a.q
      a.p < a.q  -- compute a.p in visit 0
                 -- compute a.q in visit 1
----------------------------------------------

Notes:
(1) You can define cycles this way that are not spotted by the cycle detection (because it only considers a specific subgraph of the real graph). For example, add the dependency: a.q < a.p, and you'll notice that the AG-system will crash saying that it cannot find some value in some map. There is probably some reason why the cycle detection does not consider all edges (and thus not detect all possible cycles), so will have to look at this later (perhaps together with Jeroen).
(2) You can only use defined attributes when specifying a dependency
(3) The left-hand side of a dependency declaration must be an input of the production (a synthesized attribute of a child, or a local, or an inherited attribute).