Added some syntax to globally specify types of local attributes.
If you want type signatures for the semantic functions, then the visit-sequence code generation requires 
a type signature for a local attribute if its value is required in a subsequent visit.

Often (at least in the case of EHC), a local attribute with the same name appears at multiple places 
(possibly with a different RHS, but with the same type, i.e. some naming convention for local 
attributes). So, for such local attributes it is annoying to have to specify the type signature at each 
and every location where it occurs.

Therefore a bit of a crude hack:

SIGNATURES
  ty : Ty
  ki : Ty
  knTy : Ty
  uniq1 : UID
  uniq2 : UID
  uniq3 : UID

Now I just have to specify the types of 182 local attributes of EHC...