Allow polymorphic abstract syntax (data types with type parameters).
To keep things relatively simple, the type variables can only show up in the type of a terminal, or in an argument type of a nonterminal.

For example, the following abstract syntax for expressions is parameterized by the type of variable names:

> DATA Expr v
>   | Var   nm : {v}
>   | App   f : (Expr {v})   a : (Expr {v})

> DATA Root
>   | Root  expr : (Expr {String})

The variables can only be used in a haskell type, hence the curly braces. Nonterminals can be parameterized with a list of haskell types, but then you have to wrap parentheses around the type for disambiguation purposes.


Wrappers have the appropriate types:

> WRAPPER Expr
(wrap_Expr :: T_Expr v -> Inh_Expr v -> Syn_Expr v)


You can then define an attribute |ftv| to represent the list of variables:

> ATTR Expr [ | | ftv USE {++} {[]} : {[@v]} ]

Occurrences of variables in attribute types must be prefixed with an @-symbol! (and only occurrences in attribute types, not in other types)
The reason for this is because attribute types show up at two places:
  * In attribute types of the LHS of the production:
      
> sem_Expr_Var (nm_ :: v)  =
>   let _lhsOftv :: ([v ])
>        _lhsOftv = [nm_]
>   ...

    Variables stay the same.
  
  * In attribute types of a child of the production:
  
> sem_Root_Root :: T_Expr String -> [String]
> sem_Root_Root (expr_ :: [String]) =
>   let _exprIftv :: [String]
>       _exprIftv) = expr_
>   ...

    |exprIftv| is the translation of |@expr.ftv|. Since |@expr| is of the type |Expr String|, the |@v| in the type |[@v]| is substituted by |String|.


And interface with it in the usual way:

> expr :: Expr String
> expr
>  = Expr_App (Expr_Var "id") (Expr_Var "x")

> main :: IO ()
> main
>   = let inh = Inh_Root {}
>        sem = sem_Root (Root_Root expr)
>        syn = wrap_Root sem inh
>    in putStrLn ("ftvs: " ++ show (ftv_Syn_Root syn))


Correct use of these variables is checked by the kind inferencer of the Haskell compiler.

The code generation of the AG system is slightly affected by this feature, but changes will only show up at places where polymorphic abstract syntax is used. At other places, the result will be exactly the same as with the previous version of the AG system.

Feature tests:
* Type synonyms with parameters
* Let/Case, multiple visits
* More than one type parameter
* Copy rules
* Collection rules

Compatibility test:
* The AG System
* EHC 100

To do:
* Allow a qualifier on these type variables, because you often want |Eq|, |Show|, |PP| or |Ord| on these variables in order to put them in sets or maps, or to show them for debugging purposes.
* Maybe come up with a better syntax...