New AG feature: syntax for unique numbers
Prerequisite: a threaded counter

> ATTR Tree [ | counter : Int | ]

Can be of any type that you want, with any name that you want, and you 
can pass it through the tree in whatever way you want, except that it 
MUST be a CHAINED attribute at those nonterminals whose constructors use 
the special syntax below.

Usage: using special type signatures

> SEM Tree
>  | Node
>      loc.num1 : UNIQUEREF counter
>      loc.num2 : UNIQUEREF counter

Now you have two local attributes that will have a unique number (in 
the identity-space of counter, and with the same type as counter). So, 
you can then just use the local attributes:

> SEM Tree
>  | Node 
>      loc.ids = [ @loc.num1, @loc.num2 ]

Is this all? No, there is yet one prerequisite. You must define a 
function |nextUnique|, which gets a value with the same type as the 
counter, and returns a tuple containing the next counter and one unique 
number. For example, for |Int|:

> nextUnique :: Int -> (Int, Int)
> nextUnique u = (u+1, u)

This function must be in scope of the semantic functions. The type does 
not have to be |Int|, it can be something more complicated. The name of 
this function can be changed using the --uniquedispenser=funname 
commandline option.