New AG feature!
Problem description:
Suppose you have a synthesized attribute x and want
to apply a function f to it before it is passed to
the parent. You then write:

SEM Nonterm
  | Prod
      loc.x_prev = {- previous lhs.x value -}
      lhs.x = f @loc.x_prev

There are two problems:
1) the previous lhs.x value is often computed by a
   copy-rule, use-rule, or SELF rule, and is not
   available as an attribute.
2) suppose you want to apply functions f and g
   in some arbitrary order. You then must compose
   f and g manually.

Solution:
You may now write as rule:

  AUGMENT <name of syn-attr> = <rhs>

I.e.:

SEM Nonterm
  | Production
      lhs.x = 1
      AUGMENT x = (+1)
      AUGMENT x = (+2)

The <rhs> may refer to attributes, which can
influence the evaluation order.

The order in which the aguments' are applied is
not defined. I think it's the order of appearance
in the file, but its not unlikely to be exactly
the other way around.

Typical usage:
threaded attribute you want to append to in arbitrary
order. I.e. lists of error messages, but also for
many cases where we nowadays use the USE-rules.

Similar to the "tell" operation of the writer monad.
Or to inserts in Java collections.