Added an improvement of aspect-orientedness of the AG-code, by means of a new rule:
----------
SEM MyData
  | Constr
      AROUND child = \semOfChild -> modifiedSemOfChild
----------

The AROUND statement transforms the semantics of a
child. For example, you can add additional
constructors on top of a child:

----------
DATA MyData
  | Con     body : MyData
  | Filter  body : MyData

SEM MyData
  | Con     AROUND body = \b -> sem_MyData_Filter b
  | Filter  ...
----------

We can see these extra constructors as filter nodes. With
such filtern nodes, you have a relatively easy mechanism
to conditionally override several inherited and synthesized
attributes of a child, while still having access to the
values computed by the original rules (which may be copy
rules).

The right hand side of an AROUND rule may contain
attributes. When cycle checking is enabled, such
an attribute may not depend on a synthesized attribute
of the child on the left hand side of the rule.

I don't have a good killer application for this feature
yet. I remember wanting this feature once in a while
(typically for desugaring purposes, and preventing
code duplication). For now, this addition is just a way
to have something to mention in the Haskell Communities
and Activities report.

What you cannot do yet, but what perhaps would be nice, is
an AROUND lhs, that transforms the current semantics. It
would be a mechanism to create conditional attribute
grammars
(potentially circular).