Slight parser code cleanup.
A year or longer ago, we passed options to the
parser, in order to choose between parsing double
colons and single colons, upper case tokens and lower case
tokens, square brackets and inh/syn keywords, etc.
Passing these options as a parameter was not a good idea
though, because it destroyed sharing. This you mainly
notice when you have a syntax error in a big .AG file.

So, I took some time today undoing that change, and
modified the parsers to the following structure:

parseAG = ...
  where
    opts = 
    pAg = ... <$> ... <*> ...
    ... parsers depending on opts ...
    pTypeColons
      = if doubleColons opts
        then pKey "::"
        else pKey ":"
    ... more parsers ...

For those people that use the Haskell syntax, I've
added the CHN keyword for chained attributes. Up till
now you only had INH and SYN.