Bugfix in unique number generation. Unlikely corner-case caused duplication of what should have been unique numbers.
The unique number mechanism gets an lhs.counter
and increments it for each unique number needed
in the production. The updated counter must then
be assigned to the first child of the production,
or to lhs.counter if there is no such child.

However, the mechanism did not take higher-order
children into account. Due to a funny interplay
with copy rules, this only causes problems when
there is no normal child and only a higher-order
child.