Added flags: --checkParseRhs      hs parse rhs of rules --checkParseTys      hs parse attr types --checkParseBlocks   hs parse hs blocks --checkParseHaskell  combination of the above three
Column information in the error message for
types is not always precise due to some
implementation detail (the begin pos is not
stored in the internal representation of a
type and it is too much hassle to add it).

However, if you write:
ATTR N [ | | x : {)} ]
You get an error pointing you to just slightly
after the colon, which is significantly better
than the error you would otherwise get from GHC
which is in terms of the generated HS code and
not always easily traceable back to the source.