Wanted to add a section about reference attributes to my AG section in my thesis. But also wanted to only mention things that are actually implemented in UUAG. Solution: Added "reference attributes" (Hedin-style) to the lazy evaluated version of UUAG.
For some nonterminal N
under the condition that it has one visit
and that wrappers for N are enabled
you get a constant "___node" which you may use in rules of
productions of N
and it contains a record with the synthesized attributes
of N. These you may collect in some attribute and pass around.

This is a nice feature to bridge def/use sites, and to express computations on DAGs.
Otherwise it's not really useful. But not hurting the implementation either, so it was nice to try out.