Added a modification of the kennedywarren code generation that produces an evaluator that invokes independent visits to children in parallel (using forkIO).
The option "--parallel" enables this option.

Our kennedywarren implementation attepts to infer small independent
visits to children. However, it does not try to find out how much
work is involved with a visit. When this option is enabled, forkIO
is used on all of these occasions, so it is likely that if a visit
is not a lot of work that the overhead of forking even lightweight
Haskell threads is higher than running the calculation with a
single thread.

However, the way it is organized now makes it possibly to incorporate
all sorts of parallelization heuristics - which I'm not going to
do (now) - although I'm actually curious about the performance of
the current "parallelize every visit strategy" on, say, some 32 core machine
for some example programs.