Fixed an issue where duplicate rules where produced by the automatic threading for the "uniqueRef" feature and explicitly given rules.
Explicitly given rules take precedence over the implicitly generated rules.
The implementation, however, did not do this. With this fix, the implicit
rules are not generated in this case.
This bug has been in since the beginning, but apparently only surfaced after
changes we made about half a year ago.

This commit fixes the problem reported by Jeroen Weijers. My changes should
not break anything else, but as usual some testing would be required.