UUAGC for OCaml.
E.g.:
./dist/build/uuagc/uuagc --ocaml --kennedywarren --datarecords -aO test/TestOkOcaml5.ag
ocaml test/TestOkOcaml5.ml

* Basically "all" features implemented, e.g.:
  + copy rules, SELF rules, collection rules
  + higher-order attributes, type variables
  + wrappers, builtin types (tuples, lists, maps, sets)

* Strict evaluation only, requires "absolutely non-circular" grammars.
Works for most AGs because these are usually evaluable with multiple passes.

* Purely functional implementation. Needs no special language features, except
  for rank 2 types (polymorphic record fields in OCaml).
  
This code is using a lot of continuations (similarly to how I modelled code for
step-wise evaluation before). I managed this way to get around the need for GADTs
because OCaml does not support these. This was quite some work - perhaps we should
apply these changes to the Haskell code generation too.

Some pecularities.
The order of the code blocks in the .AG file matters, because the order of
let-bindings in OCaml matters. In general, the code blocks will be put in
front of the generated code, meaning that code in blocks cannot refer to
the generated functions. With some block-annotatiosn, you can put code blocks
after the generated code, which can thus refer to the generated code
(but not the other way around). I've also organized the generated code into
a module for data types and one for code, which perhaps makes it easier to
determine what is exported at top-level.

I didn't put any support for multi-core concurrency in, because OCaml does not
provide facilities for this, at least, as far as I know. There were apparently
some projects in the past. Anyway, for purely-functional sequential code, it
should actually be pretty efficient.

There are likely still many slight code generation bugs here and there,
but it should already work pretty well.
See test/TestOkOcaml*.ag for examples.

***
My changes should not have affected the code generation for
Haskell, but practice learns that making some changes usually breaks some
couple of things.