Wanted to try out a happstack application, and made an uuagc-as-a-webservice.
To try out, install uuagc and uuagc-web from the repo (they are not on hackage yet).
Then run the server with:

  uuagc-web -p 2000

Open http://localhost:2000 and upload your ag source file to get a haskell source.
(be sure to put commandline options like -a in the options-field otherwise you will
not get a lot of output).

Or, use the shell script "uuagc.sh" in the frontends directory:

  ./frontends/uuagc.sh localhost:2000 -aO myagfile.ag

It would be nice to put this online somewhere during the AG tutorial at ICFP. Then
people can experiment with it without actually having the need to install it.
Certainly, when you have the haskell platform with cabal and everything, you more
easily install the uuagc system through cabal...
but if you only have a bare GHC or Ocaml...

Making the happstack application was a piece of cake. Writing a portable shell script on the other hand :X