Configure flag for bootstrapping uuagc with some 'uuagc' in $PATH.
The default behavior is bootstrapping with the 'uuagc-bootstrap' cabal
package. If you don't want to use that package but want to use some
other uuagc executable instead, you can run:
> cabal configure --ghc-options=-DEXTERAL_UUAGC

This is a bit complicated, but I don't see an easier way.
It seems a deficiency of Cabal that you cannot write a build
configuration for the Setup.hs...